pyproject.toml
[project]
name = "forecasting_sa"
version = "0.0.1"
dependencies = [
    "rpy2",
    "sktime",
    "gluonts",
    "pytorch-lightning",
    "mxnet",
    "kaleido",
    "Jinja2",
    "omegaconf",
    "statsforecast",
    "missingno",
]

[project.optional-dependencies]
tests = [
    "setuptools",
    "wheel",
    "pyyaml",
    "dbx",
    "pytest",
    "pytest-cov",
    "pyspark==3.3.0",
    "PyArrow>=1.0.0",
    "sqlalchemy",
    "alembic",
    "hyperopt",
    "plotly",
    "black",
    "datasetsforecast"
]

[tool.setuptools.packages.find]
include = ["forecasting_sa*"]
exclude = ["tests*"]
#namespaces = false
